+++ src/main/java/chess/infra/connection/ConnectionGenerator.java
@@ -1,4 +1,4 @@
-package chess.infra;
+package chess.infra.connection;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
